# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateRequest": {
        "additionalProperties": false,
        "properties": {
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "spec": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Spec",
            "type": "object"
          }
        },
        "required": [
          "project_id",
          "spec"
        ],
        "title": "CreateRequest",
        "type": "object"
      },
...
      "FeatureBinding": {
        "additionalProperties": false,
        "properties": {
          "asset_id": {
            "maxLength": 256,
            "minLength": 1,
            "pattern": ".*\\S.*",
            "title": "Asset Id",
            "type": "string"
          },
          "band": {
            "default": 1,
            "minimum": 1.0,
            "title": "Band",
            "type": "integer"
          },
          "name": {
            "maxLength": 256,
            "minLength": 1,
            "pattern": ".*\\S.*",
            "title": "Name",
            "type": "string"
          },
          "unit": {
            "maxLength": 256,
            "minLength": 1,
            "pattern": ".*\\S.*",
            "title": "Unit",
            "type": "string"
          },
          "version": {
            "minimum": 1.0,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "unit",
          "asset_id",
          "version"
        ],
        "title": "FeatureBinding",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JsonValue": {},
...
      "PrepareFrameRequest": {
        "additionalProperties": false,
        "properties": {
          "features": {
            "items": {
              "$ref": "#/components/schemas/FeatureBinding"
            },
            "maxItems": 8,
            "minItems": 1,
            "title": "Features",
            "type": "array"
          },
          "name": {
            "maxLength": 256,
            "minLength": 1,
            "pattern": ".*\\S.*",
            "title": "Name",
            "type": "string"
          },
          "options": {
            "$ref": "#/components/schemas/RasterFrameOptions"
          },
          "period": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TimeRange"
              },
              {
                "type": "null"
              }
            ]
          },
          "project_id": {
            "maxLength": 256,
            "minLength": 1,
            "pattern": ".*\\S.*",
            "title": "Project Id",
            "type": "string"
          },
          "time_resolution": {
            "anyOf": [
              {
                "maxLength": 256,
                "minLength": 1,
                "pattern": ".*\\S.*",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Resolution"
          }
        },
        "required": [
          "project_id",
          "name",
          "features",
          "options"
        ],
        "title": "PrepareFrameRequest",
        "type": "object"
      },
...
      "RasterFrameOptions": {
        "additionalProperties": false,
        "properties": {
          "mode": {
            "enum": [
              "all",
              "sample"
            ],
            "title": "Mode",
            "type": "string"
          },
          "response_index": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response Index"
          },
          "sample_size": {
            "default": 1000,
            "maximum": 10000.0,
            "minimum": 4.0,
            "title": "Sample Size",
            "type": "integer"
          },
          "seed": {
            "default": 42,
            "maximum": 2147483647.0,
            "minimum": 0.0,
            "title": "Seed",
            "type": "integer"
          },
          "standardize": {
            "title": "Standardize",
            "type": "boolean"
          }
        },
        "required": [
          "mode",
          "standardize"
        ],
        "title": "RasterFrameOptions",
        "type": "object"
      },
...
      "TimeRange": {
        "additionalProperties": false,
        "properties": {
          "end": {
            "format": "date-time",
            "title": "End",
            "type": "string"
          },
          "start": {
            "format": "date-time",
            "title": "Start",
            "type": "string"
          }
        },
        "required": [
          "start",
          "end"
        ],
        "title": "TimeRange",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "CoastMAS",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/data-assets": {
      "get": {
        "operationId": "list_data_api_v1_data_assets_get",
        "parameters": [
          {
            "in": "query",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": {
                      "$ref": "#/components/schemas/JsonValue"
                    },
                    "type": "object"
                  },
                  "title": "Response List Data Api V1 Data Assets Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Data"
      },
      "post": {
        "operationId": "create_data_api_v1_data_assets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "$ref": "#/components/schemas/JsonValue"
                  },
                  "title": "Response Create Data Api V1 Data Assets Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Data"
      }
    },
...
    "/api/v1/data-assets/prepare-raster-frame": {
      "post": {
        "operationId": "prepare_api_v1_data_assets_prepare_raster_frame_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrepareFrameRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "$ref": "#/components/schemas/JsonValue"
                  },
                  "title": "Response Prepare Api V1 Data Assets Prepare Raster Frame Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Prepare",
        "tags": [
          "data"
        ]
      }
    },
...
  }
}